vc4: Use a pool of CMA for vc4_bo allocations
When allocating a vc4_bo, instead of directly allocating CMA memory for
it, a single 128MB pool of CMA memory is allocated and then the memory
for the bo is just set as an offset into that buffer. The intention is
to avoid a situation where user-space can eat up all of the CMA memory.

When the pool is full and it can’t find a slot to insert a new buffer it
will first try purgeing all of the purgeable buffers as usual. If that
still doesn’t free up a large enough slot it will try paging out
buffers. This means that it will make a copy of the buffer contents into
vmalloc’d memory and free up the slot. The buffers are maintained in a
linked list in most-recently used order so that it can prioritise paging
out buffers that haven’t been recently used.

When a buffer is used by increasing the usecnt it is automatically paged
back in to the CMA pool from the vmalloc’d copy. The same also happens
in the page fault handler for when the buffer is accessed via the
user-space mapping.

The paging algorithm avoids paging out any buffers that are marked with
a usecnt. It also explicitly avoids ever paging out the console
framebuffer and the binning buffer because these are used internally in
the kernel without increasing the usecnt.

The buffer caching mechanism has been removed under the assumption that
allocating buffers no longer entails an actual CMA allocation, so it
shouldn’t be as slow.

To implement the pool, vc4_bo has been changed to directly inherit from
drm_gem_object instead of using dma_gem_cma_object. In order to make
this work, the patch replaces any parts of the code that were previously
accessing the paddr or vaddr of dma_gem_cma_object with a helper
function that calculates them using the buffer’s offset and the base
addresses of the CMA pool buffer.