vc4_bo: Make the CMA pool offset list use a special node struct
Instead of the CMA pool directly using a linked list of vc4_bo’s to
maintain the list of addresses, it is now a list of vc4_offset_node.
This is a little struct that just contains the offset and the size.
Later patches will take advantage of this to make the list contain
things other than a vc4_bo as well. The node is directly embedded in a
vc4_bo, so when the node represents a vc4_bo (ie, currently always),
then the address of the bo can be obtained with container_of.