vc4: Add a param for user-space to query CMA pool size
User-space can now query the CMA pool size via
DRM_VC4_PARAM_MEMORY_POOL_SIZE. This reports the size of the CMA pool
minus the size of the buffers that can’t be paged out. The intention is
for Mesa to use this to try to avoid submitting jobs that will require
more than this amount of memory.