vc4/bo: Set up the user-space mapping while holding the BO lock
When handling a page fault for a buffer that is paged out of the CMA
pool, it now sets up the DMA mapping while still holding the buffer
lock. Otherwise there is a chance that something will cause the buffer
to be paged out before the mapping is set up. In that case the mapping
would point to the wrong area of the CMA pool and it wouldn’t cause a
page fault when accessed.