vc4/bo: Don’t use the gem_create_object callback
It looks like this callback is only meant to be used to let the CMA
helper allocate the object. Since the CMA pool was added, the CMA helper
is no longer used so the callback is no longer needed. The main
advantage of this is that the type label can be set on the object
immediately and there is no need for the dummy KERNEL type. This is
important now that we are trying to avoid paging out certain types of
objects (such as BCL and RCL). Without this patch there would be a small
window where the buffer could be paged out before it gets the right
type.