vc4/bo: Mark the buffer as in use when validating a shader
After creating a buffer for a shader there was nothing stopping the
driver from paging out the buffer while the user data is being copied in
and the shader gets validated. This patch makes it set the usecnt on the
buffer during that time to keep it in CMA memory and make sure it has a
valid vaddr.