vc4: Use a pool of CMA for vc4_bo allocations
Instead of allocating CMA memory for each vc4_bo, there is now a single
pool of 128MB of CMA memory. The memory for a buffer is now normally
stored using the SHMEM helpers instead. Whenever the buffer needs to be
used by the hardware it will be copied into a region of the CMA pool.
The intention is to avoid a situation where user-space can eat up all of
the CMA memory.

When the pool is full and it can’t find a slot to insert a new buffer it
will first try purgeing all of the purgeable buffers as usual. If that
still doesn’t free up a large enough slot it will try paging out
buffers. This means that the buffer will give up its space in the CMA
pool and copy the contents to the SHMEM backing if it has been used by
the hardware for rendering. The buffers in the pool are maintained in a
linked list in most-recently used order so that it can prioritise paging
out buffers that haven’t been recently used.

When a buffer is used by increasing the usecnt it is automatically paged
back in to the CMA pool and its contents copied if need be. The driver
detects that a buffer may have been modified by userspace by setting a
flag in the page fault handler. Whenever the buffer is copied into the
CMA pool the user-space mappings for the buffer are invalidated so that
we can detect any further changes as well.

The paging algorithm avoids paging out any buffers that are marked with
a usecnt. It also explicitly avoids ever paging out the console
framebuffer and the binning buffer because these are used internally in
the kernel without increasing the usecnt.

To implement the pool, vc4_bo has been rebased to inherit from
drm_gem_shmem_object instead of using dma_gem_cma_object. In order to
make this work, the patch replaces any parts of the code that were
previously accessing the paddr or vaddr of dma_gem_cma_object with a
helper function that calculates them using the buffer’s offset and the
base addresses of the CMA pool buffer.