drm/shmem-helper: Don't adjust the vm_area_struct pgoff
When mmapping the shmem, it would previously adjust the pgoff in the
vm_area_struct to remove the fake offset that is added to be able to
identify the buffer. This patch removes the adjustment and makes the
fault handler use the vm_fault address to calculate the page offset
instead. Although using this address is apparently discouraged, several
other DRM drivers also seem to be doing it.

The problem with removing the pgoff is that it prevents
drm_vma_node_unmap from working because that searches the mapping tree
by address but that doesn't work because all of the mappings are at
offset 0. drm_vma_node_unmap is being used by the shmem helpers when
purging the buffer.