vc4/bo: Move the init code for vc4_dev.purgeable to vc4_bo
Previously the init code for the purgeable struct inside vc4_dev was
handled in vc4_gem_init. This was a bit odd because the rest of the code
that touches this struct lives in vc4_bo.c. This patch adds a separate
function to initialize it in a similar way to vc4_bo_cache_init. This
will help to be able to add more initialisation later that uses
static functions internal to vc4_bo.c.