vc4/bo: Use a range for the dirty region of the shmem backing
Instead of using a single boolean to mark whether the shmem backing
needs to be copied into the CMA pool, it now uses a range covering the
minimum and maximum pages that have had a fault. That way we can reduce
the amount of memory copied if user-space is using a large buffer and
updating only small subregions of it.